made separate components for DeleteModal and Navbar
